Refactor + Comment
