+++ src/components/Card/ProcessCard/Styles.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components/macro";
-import {CardContainer} from "../Card/Styles";
+import {CardContainer} from "../Styles";
 
 export const CardProcess = styled(CardContainer)`
   width: 500px;
